requirements.txt
requests
flask
mysql-connector-python
bs4
lxml

